+++ CHANGELOG.md
@@ -21,6 +21,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 * Entity variables in tests don't start with `fake` anymore
 * Moved `BasePaginationParameters` and `Exceptions` and `ValueObject` to api project
 
+### Fixed
+
+* Can handle no global git config
+
 ## [0.22.1] - 08/13/2023
 
 ### Fixed